index.js and style.css
in this project, when i go through the page the toggle swith is in dark mode but the website is in light mode. after one clicking toggle switch it will not change the theme.and again it is in initial position it will change it will become dark mode. i fixed it.